+++ src/geometry/rtree.h
@@ -5,43 +5,22 @@
 #include <boost/geometry.hpp>
 #include <boost/geometry/geometries/geometries.hpp>
 #include <boost/geometry/index/rtree.hpp>
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
-#include <unordered_map>
-
-#include "exceptions.h"
-#include "utilities.h"
-#include <memory>
-#include <opencv2/imgproc.hpp>
-#include <opencv2/opencv.hpp>
-#include <pybind11/numpy.h>
-#include <pybind11/pybind11.h>
-#include <stdexcept>
-#include <string>
 #include <unordered_map>
 #include <vector>
 
-// #define DLUPDEBUG
-
 namespace bg = boost::geometry;
 namespace bgi = boost::geometry::index;
-namespace py = pybind11;
 
 using BoostPoint = bg::model::d2::point_xy<double>;
-using BoostPolygon = bg::model::polygon<BoostPoint>;
 using BoostBox = bg::model::box<BoostPoint>;
-using BoostRing = bg::model::ring<BoostPoint>;
-using BoostLineString = bg::model::linestring<BoostPoint>;
-using BoostMultiPolygon = bg::model::multi_polygon<BoostPolygon>;
 
-class GeometryCollection; // Forward declaration
-
-class RTreeWrapper {
+class RTreeBase {
 public:
     using RTreeType = bgi::rtree<std::pair<BoostBox, size_t>, bgi::quadratic<16>>;
 
-    RTreeWrapper(GeometryCollection *geometryCollection)
-        : rTreeInvalidated(true), geometryCollection(geometryCollection) {}
+    virtual ~RTreeBase() = default;
+
+    virtual void rebuild() = 0; // Pure virtual function for rebuilding the R-tree
 
     void insert(const BoostBox &box, size_t index) {
         rtree.insert(std::make_pair(box, index));
@@ -65,14 +44,9 @@ public:
 
     bool isInvalidated() const { return rTreeInvalidated; }
 
-    void rebuild();
-
-private:
+protected:
     RTreeType rtree;
-    bool rTreeInvalidated;
-    GeometryCollection *geometryCollection; // Pointer to GeometryCollection
+    bool rTreeInvalidated = true;
 };
-
-
 
 #endif // DLUP_GEOMETRY_RTREE_H
